+++ lib/rubocop/ast.rb
@@ -5,6 +5,7 @@ require 'forwardable'
 require 'set'
 
 require_relative 'ast/ext/range'
+require_relative 'ast/ext/set'
 require_relative 'ast/node_pattern'
 require_relative 'ast/sexp'
 require_relative 'ast/node'